ajout symetrie Cercle

+++ src/main/patchwork/Point.java
@@ -21,7 +21,7 @@ public class Point {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
